#13: Cart http://pm.5psolutions.net/public/index.php?path_info=projects/training-cosmin-barbascu/tasks/13 Changed float to decimal and added unsigned and index to foreign keys

+++ database/migrations/2020_09_04_101136_create_products_table.php
@@ -17,7 +17,7 @@ class CreateProductsTable extends Migration
             $table->id();
             $table->string('title');
             $table->string('description');
-            $table->float('price');
+            $table->decimal('price');
         });
     }
